src/main/de/hsmainz/cs/semgis/arqextension/linestring/attribute/StartPoint.java
/** *****************************************************************************
 * Copyright (c) 2017 Timo Homburg, i3Mainz.
 * All rights reserved. This program and the accompanying materials
 * are made available under the terms of the BSD License
 * which accompanies this distribution, and is available at
 * https://directory.fsf.org/wiki/License:BSD_4Clause
 *
 * This project extends work by Ian Simmons who developed the Parliament Triple Store.
 * http://parliament.semwebcentral.org and published his work und BSD License as well.
 *
 *
 ****************************************************************************** */
package de.hsmainz.cs.semgis.arqextension.linestring.attribute;

import org.apache.jena.sparql.expr.NodeValue;
import org.locationtech.jts.geom.Geometry;

import io.github.galbiston.geosparql_jena.implementation.GeometryWrapper; import io.github.galbiston.geosparql_jena.implementation.GeometryWrapperFactory;
import org.apache.jena.datatypes.DatatypeFormatException;
import org.apache.jena.sparql.expr.ExprEvalException;
import org.apache.jena.sparql.function.FunctionBase1;
import org.locationtech.jts.geom.LineString;
import org.locationtech.jts.geom.Point;

public class StartPoint extends FunctionBase1 {

    @Override
    public NodeValue exec(NodeValue arg0) {

        try {
            GeometryWrapper geometry = GeometryWrapper.extract(arg0);
            Geometry geom = geometry.getXYGeometry();

            if (geom instanceof LineString) {

                Point point = ((LineString) geom).getStartPoint();
                GeometryWrapper pointWrapper = GeometryWrapperFactory.createGeometry(point, geometry.getSrsURI(), geometry.getGeometryDatatypeURI());
                return pointWrapper.asNodeValue();
            }

            throw new ExprEvalException("Input geometry is not a LineString");
        } catch (DatatypeFormatException ex) {
            throw new ExprEvalException(ex.getMessage(), ex);
        }
    }

}
